wallet: move coin selection, key lookup, and locking to managers
Move specialized logic from wallet.go to respective managers:
- Coin and selection strategy to tx_creator.go.
- PrivKeyForAddress to signer.go.
- LockedOutpoint to utxo_manager.go.

This further decouples the Wallet struct and aligns with the manager-based
architecture.
